use unsigned long for memory-related quantitites. Helps on 64-bit machines and in the 2-4Gb region
git-svn-id: https://svn.r-project.org/R/trunk@20094 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/include/Startup.h
@@ -54,10 +54,10 @@ typedef struct
     Rboolean DebugInitFile;
     SA_TYPE RestoreAction;
     SA_TYPE SaveAction;
-    int vsize;
-    int nsize;
-    int max_vsize;
-    int max_nsize;
+    unsigned long vsize;
+    unsigned long nsize;
+    unsigned long max_vsize;
+    unsigned long max_nsize;
     int NoRenviron;
 
       /* Permanent copy of the command line arguments and the number
